requirements.txt
requests
pypact-lang
paho-mqtt